refactor: overhaul builders structure
