+++ src/builders/rollup/config.ts
@@ -1,0 +1,146 @@
+import type { OutputOptions, PreRenderedChunk } from "rollup";
+import commonjs from "@rollup/plugin-commonjs";
+import { nodeResolve } from "@rollup/plugin-node-resolve";
+import alias from "@rollup/plugin-alias";
+import replace from "@rollup/plugin-replace";
+import { resolve, isAbsolute } from "pathe";
+import { arrayIncludes, getpkg, warn } from "../../utils";
+import type { BuildContext, RollupOptions } from "../../types";
+import { esbuild } from "./plugins/esbuild";
+import { JSONPlugin } from "./plugins/json";
+import { rawPlugin } from "./plugins/raw";
+import { cjsPlugin } from "./plugins/cjs";
+import { shebangPlugin } from "./plugins/shebang";
+import {
+  DEFAULT_EXTENSIONS,
+  getChunkFilename,
+  resolveAlias,
+  resolveAliases,
+} from "./utils";
+
+export function getRollupOptions(ctx: BuildContext): RollupOptions {
+  const _aliases = resolveAliases(ctx);
+  return (<RollupOptions>{
+    input: Object.fromEntries(
+      ctx.options.entries
+        .filter((entry) => entry.builder === "rollup")
+        .map((entry) => [
+          entry.name,
+          resolve(ctx.options.rootDir, entry.input),
+        ]),
+    ),
+
+    output: [
+      ctx.options.rollup.emitCJS &&
+        <OutputOptions>{
+          dir: resolve(ctx.options.rootDir, ctx.options.outDir),
+          entryFileNames: "[name].cjs",
+          chunkFileNames: (chunk: PreRenderedChunk) =>
+            getChunkFilename(ctx, chunk, "cjs"),
+          format: "cjs",
+          exports: "auto",
+          interop: "compat",
+          generatedCode: { constBindings: true },
+          externalLiveBindings: false,
+          freeze: false,
+          sourcemap: ctx.options.sourcemap,
+          ...ctx.options.rollup.output,
+        },
+      <OutputOptions>{
+        dir: resolve(ctx.options.rootDir, ctx.options.outDir),
+        entryFileNames: "[name].mjs",
+        chunkFileNames: (chunk: PreRenderedChunk) =>
+          getChunkFilename(ctx, chunk, "mjs"),
+        format: "esm",
+        exports: "auto",
+        generatedCode: { constBindings: true },
+        externalLiveBindings: false,
+        freeze: false,
+        sourcemap: ctx.options.sourcemap,
+        ...ctx.options.rollup.output,
+      },
+    ].filter(Boolean),
+
+    external(id) {
+      id = resolveAlias(id, _aliases);
+      const pkg = getpkg(id);
+      const isExplicitExternal =
+        arrayIncludes(ctx.options.externals, pkg) ||
+        arrayIncludes(ctx.options.externals, id);
+      if (isExplicitExternal) {
+        return true;
+      }
+      if (
+        ctx.options.rollup.inlineDependencies ||
+        id[0] === "." ||
+        isAbsolute(id) ||
+        /src[/\\]/.test(id) ||
+        id.startsWith(ctx.pkg.name!)
+      ) {
+        return false;
+      }
+      if (!isExplicitExternal) {
+        warn(ctx, `Inlined implicit external ${id}`);
+      }
+      return isExplicitExternal;
+    },
+
+    onwarn(warning, rollupWarn) {
+      if (!warning.code || !["CIRCULAR_DEPENDENCY"].includes(warning.code)) {
+        rollupWarn(warning);
+      }
+    },
+
+    plugins: [
+      ctx.options.rollup.replace &&
+        replace({
+          ...ctx.options.rollup.replace,
+          values: {
+            ...ctx.options.replace,
+            ...ctx.options.rollup.replace.values,
+          },
+        }),
+
+      ctx.options.rollup.alias &&
+        alias({
+          ...ctx.options.rollup.alias,
+          entries: _aliases,
+        }),
+
+      ctx.options.rollup.resolve &&
+        nodeResolve({
+          extensions: DEFAULT_EXTENSIONS,
+          ...ctx.options.rollup.resolve,
+        }),
+
+      ctx.options.rollup.json &&
+        JSONPlugin({
+          ...ctx.options.rollup.json,
+        }),
+
+      shebangPlugin(),
+
+      ctx.options.rollup.esbuild &&
+        esbuild({
+          sourcemap: ctx.options.sourcemap,
+          ...ctx.options.rollup.esbuild,
+        }),
+
+      ctx.options.rollup.commonjs &&
+        commonjs({
+          extensions: DEFAULT_EXTENSIONS,
+          ...ctx.options.rollup.commonjs,
+        }),
+
+      ctx.options.rollup.preserveDynamicImports && {
+        renderDynamicImport() {
+          return { left: "import(", right: ")" };
+        },
+      },
+
+      ctx.options.rollup.cjsBridge && cjsPlugin({}),
+
+      rawPlugin(),
+    ].filter(Boolean),
+  }) as RollupOptions;
+}
