+++ src/builders/rollup/index.ts
@@ -1,0 +1,1 @@
+export { rollupBuild } from "./build";
