+++ src/builders/rollup/plugins/shebang.ts
